Custom LCD Library
This is a low level driver for an LCD using a Hitachi HD44780 or
compatible chipset.  This library is based on the Arduino's Liquid
Crystal library, except this provides low level functionality instead of
high level functionality.  In addition, this library allows you to read
as well as write.

This class also does not have hard coded waits, but rather checks the
busy flag (which the Liquid Crystal library does not).  As a result,
this library runs slightly faster as a result.